Deploy feature to propagate config refresh across several instances of the microservice
This feature is implemented with kafka as an event system. Microservices
deployed on multiple VMs or pods are load balanced, so the /refresh http
call only hits one instance at a time. When this request hits, the
instance handling the request logs a refresh event to the shared kafka
topic. Other instances of the microservice listen for events on the
shared topic and pull config from Config Server when each instance
recieves the event. All instances are producers and consumers. Only one
instance can produces per load balanced request.